RaftActor commits index not to conflict with succeeding AppendEntries
On the receiving of AppendEntries message, except for empty entries,
the receiver RaftActor cannot commit `leaderCommit` if `leaderCommit`
is greater than or equal to the last index of received entries. Other
succeeding AppendEntries messages can conflict with an existing entry
with an index less than or equal to `leaderCommit` but greater than
the last index of received entries. Instead, the receiver can commit
the last index of received entries.