Non-leaders passivate an entity eventually
A non-leader (follower or candidate) asks permission to passivate an
entity to the leader. The leader replies with a permit or denial based
on whether the leader's entity is running. If the non-leader receives a
permit, it will passivate its entity. This mechanism ensures that
non-leaders can passivate their entity eventually after the leader
passivate the leader's entity.