Fix old snapshot read in consecutive snapshot synchronization
RaftActor of the source replica group of the synchronization can execute
compactions during snapshot synchronization. In such a case, the single
SnapshotSyncManager instance could perform multiple snapshot
synchronizations in serial order. If the SnapsotSyncManager does so, it
could read an old snapshot in the newer synchronization. This old
snapshot read is related to the following aspects:

* SnapshotSyncManager reads snapshots from a SnapshotStore that the
  SnapshotSyncManager spawns as its descendant. The SnapshotStore is
  a different instance from a SnapshotStore running as a descendant of
  RaftActor of the source replica group. The SnapshotStore cannot read
  new snapshots the other one wrote (by such as compaction) if it has
  already recovered from persisted data.
* SnapshotSyncManager reuses an instance of SnapshotStore among multiple
  snapshot synchronizations. It cannot read a new snapshot written by a
  compaction that happens during the synchronization.

Change SnapshotSyncManager such that it executes only one snapshot
synchronization in its lifecycle. That means SnapshotSyncManager doesn't
reuse SnapshotStore instances. Each synchronization will be performed by
a different SnapshotSyncManager instance and in serial order.